+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Improve `stud items:show` command to display description in dedicated sections with automatic text sanitization and divider detection [SCI-8]
 
+### Fixed
+- Remove brittle statusCategory filtering from `stud items:start` transition lookup to show all available transitions instead of filtering by 'in_progress' status category [SCI-13]
+
 ## [2.4.1] - 2025-11-25
 
 ### Removed